torch_serving/pyproject.toml
[tool.poetry]
name = "torch-serving"
version = "0.1.0"
description = "A Python-Based Example for Model Serving with PyTorch."
authors = ["Martynas Subonis <martynas.subonis@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "0.115.2"
gcloud-aio-storage = "9.3.0"
numpy = "2.1.2"
pillow = "10.4.0"
pydantic = "2.9.2"
torch = "2.3.1"
torchvision = "0.18.1"
uvicorn = "0.31.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
